move torch utilities
